+++ src/main/java/com/amigoscode/dto/CartResponse.java
@@ -1,12 +1,13 @@
 package com.amigoscode.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class CartResponse {
     private Long cartId;
     private List<CartItemResponse> items;
-    private Double totalPrice; 
+    private BigDecimal totalPrice;
     private Integer totalItems;
-} 
+}
