+++ src/main/java/com/amigoscode/dto/CouponValidateResponse.java
@@ -4,15 +4,17 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class CouponValidateResponse {
     private boolean valid;
-    private String message;         // "Áp dụng thành công!" hoặc lý do thất bại
+    private String message;         // "Áp dụng thành công!" or reason for failure
     private String code;
     private String discountType;    // "PERCENT" | "FIXED"
-    private double discountValue;   // Giá trị giảm
-    private double discountAmount;  // Số tiền thực tế được giảm (đã tính theo subtotal)
-    private double finalTotal;      // Tổng sau khi giảm
+    private BigDecimal discountValue;   // Configured discount value
+    private BigDecimal discountAmount;  // Actual money saved (computed against subtotal)
+    private BigDecimal finalTotal;      // Total after applying discount
 }
